Fix watch method find .vt crawling
