+++ src/cmd/root/watch.ts
@@ -6,6 +6,7 @@ import sdk from "~/sdk.ts";
 import { FIRST_VERSION_NUMBER, STATUS_COLORS } from "~/consts.ts";
 import { displayStatusChanges } from "~/cmd/git/utils.ts";
 import { getTotalChanges } from "~/vt/git/utils.ts";
+import { vtRootOrFalse } from "~/cmd/utils.ts";
 
 /**
  * Formats a version range string based on the first, current, and latest versions.
@@ -82,32 +83,41 @@ export const watchCmd = new Command()
     { default: 300 },
   )
   .action(async (options) => {
-    const cwd = Deno.cwd();
     const spinner = new Kia("Starting watch mode...");
-    spinner.start();
-    const vt = VTClient.from(cwd);
 
-    // Get initial branch information for display
-    const {
-      currentBranch: currentBranchId,
-      version: currentVersion,
-      projectId,
-    } = await vt.getMeta().loadConfig();
-    const currentBranch = await sdk.projects.branches.retrieve(
-      projectId,
-      currentBranchId,
-    );
+    const vtRoot = await vtRootOrFalse(spinner);
+    if (!vtRoot) return;
 
-    spinner.stop();
-    console.log(
-      `On branch ${colors.cyan(currentBranch.name)}@${
-        getVersionRangeStr(
-          FIRST_VERSION_NUMBER,
-          currentVersion,
-          currentBranch.version,
-        )
-      }`,
-    );
+    let vt: VTClient;
+    try {
+      spinner.start();
+      vt = VTClient.from(vtRoot);
+
+      // Get initial branch information for display
+      const {
+        currentBranch: currentBranchId,
+        version: currentVersion,
+        projectId,
+      } = await vt.getMeta().loadConfig();
+      const currentBranch = await sdk.projects.branches.retrieve(
+        projectId,
+        currentBranchId,
+      );
+
+      spinner.stop();
+      console.log(
+        `On branch ${colors.cyan(currentBranch.name)}@${
+          getVersionRangeStr(
+            FIRST_VERSION_NUMBER,
+            currentVersion,
+            currentBranch.version,
+          )
+        }`,
+      );
+    } finally {
+      spinner.stop();
+    }
+
     console.log("Watching for changes. Press Ctrl+C to stop.");
 
     const watchingForChangesLine = () =>
