+++ src/vt/git/clone.ts
@@ -4,7 +4,6 @@ import { removeEmptyDirs } from "~/utils.ts";
 import { shouldIgnore } from "~/vt/git/paths.ts";
 import * as path from "@std/path";
 import { ensureDir } from "@std/fs";
-import type ValTown from "@valtown/sdk";
 import { doAtomically } from "~/vt/git/utils.ts";
 
 /**
